Update submodules, fix a build issue
With the Maven Shade plugin, executing the package phase after a
successful build left output files in target/ causes Shade to process
the shaded JAR as if it was the original JAR. This causes "overlapping
classes" warnings and causes the services files to grow on each
iteration. Teach build.bash to clean the build first.